Sends the Begin token through all outputs before thread start.
We can have a concurrency problem if the transforms handle the sending
of Begin token at the same time they initialize() because if for one
reason a transform runs a lot slower than another the input can be
killed before every candidate had a chance to register. On the design
side, this change is annoying because it adds even more "data" knowledge
(or at least metadata/control flow) to the thread object.